requirements-dev.txt
# Testing Framework
pytest>=8.3.0,<9.0.0
pytest-cov>=6.0.0,<7.0.0
pytest-asyncio>=0.24.0,<1.0.0

# HTTP Testing
httpx>=0.28.0,<1.0.0
responses>=0.25.0,<1.0.0

# Configuration Testing
pyyaml>=6.0.0,<7.0.0
